fix: Manually specify the HTTP Methods to workaround a Serverless -> Cloudformation issue
We were getting the following error:

```Error: The CloudFormation template is invalid: Template format error: Resource name ApiGatewayMethodWebhook* is non alphanumeric.```

Seems like Serverless is generating some CloudFormation resource with a name that contains the HTTP Method that was specified in serverless.yml. Our workaround is to just manually specify the HTTP methods.